+++ microservice/beamable.tooling.common/OpenAPI/SchemaGenerator.cs
@@ -9,6 +9,7 @@ using Microsoft.OpenApi.Interfaces;
 using Microsoft.OpenApi.Models;
 using System.Collections;
 using System.Reflection;
+using Beamable.Common.Semantics;
 using UnityEngine;
 using ZLogger;
 using static Beamable.Common.Constants.Features.Services;
@@ -40,7 +41,11 @@ public class SchemaGenerator
 		public bool IsFromBeamGenerateSchema() => SourceCallable == null && Type.GetCustomAttribute<BeamGenerateSchemaAttribute>() != null;
 
 		public bool IsFromCallableWithNoClientGen() => IsFromCallable() && SourceCallable.Method.GetCustomAttribute<CallableAttribute>(true).Flags.HasFlag(CallableFlags.SkipGenerateClientFiles);
-		public bool ShouldNotGenerateClientCode() => (IsFromFederation() || IsFromCallableWithNoClientGen()) && !IsFromBeamGenerateSchema();
+
+		public bool IsBeamSemanticType() => Type.GetInterfaces()
+			.Any(i => i.GetGenericTypeDefinition() == typeof(IBeamSemanticType<>));
+		
+		public bool ShouldSkipClientCodeGeneration() => (IsFromFederation() || IsFromCallableWithNoClientGen() || IsBeamSemanticType()) && !IsFromBeamGenerateSchema();
 		
 		public bool IsPrimitive()
 		{
@@ -274,7 +279,17 @@ public class SchemaGenerator
 			case { } x when x.IsGenericType && x.GetGenericTypeDefinition() == typeof(Optional<>):
 				return Convert(x.GetGenericArguments()[0], ref requiredTypes,depth - 1, sanitizeGenericType);
 			case { } x when x.IsGenericType && x.GetGenericTypeDefinition() == typeof(Nullable<>):
-				return Convert(x.GetGenericArguments()[0], ref requiredTypes,depth - 1, sanitizeGenericType);
+				return Convert(x.GetGenericArguments()[0], depth - 1);
+			case { } x when x.GetInterfaces().Any(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof(IBeamSemanticType<>)):
+				var semanticType = x.GetInterfaces().First(i => i.IsGenericType && i.GetGenericTypeDefinition() == typeof(IBeamSemanticType<>));
+				var semanticTypeSchema = Convert(semanticType.GetGenericArguments()[0], depth - 1);
+				if (Activator.CreateInstance(runtimeType) is IBeamSemanticType semanticTypeInstance)
+				{
+					semanticTypeSchema.Extensions[SCHEMA_SEMANTIC_TYPE_NAME_KEY] =
+						new OpenApiString(semanticTypeInstance.SemanticName);
+				}
+
+				return semanticTypeSchema;
 			case { } x when x == typeof(double):
 				return new OpenApiSchema { Type = "number", Format = "double" };
 			case { } x when x == typeof(float):
@@ -456,6 +471,16 @@ public class SchemaGenerator
 
 					var comment = DocsLoader.GetMemberComments(member);
 					fieldSchema.Description = comment?.Summary;
+
+
+					Type classSemanticType = member.FieldType.GetInterfaces().FirstOrDefault(i =>
+						i.IsGenericType && i.GetGenericTypeDefinition() == typeof(IBeamSemanticType<>));
+					if (classSemanticType != null && Activator.CreateInstance(member.FieldType) is IBeamSemanticType memberSemanticTypeInstance)
+					{
+						fieldSchema.Extensions[SCHEMA_SEMANTIC_TYPE_NAME_KEY] =
+							new OpenApiString(memberSemanticTypeInstance.SemanticName);
+					}
+
 					schema.Properties[name] = fieldSchema;
 
 					if (!member.FieldType.IsAssignableTo(typeof(Optional)))